refactor: rename IWordCard.ts
