+++ src/application/eBookPage/ebookInterface.ts
